0.20.20-ALPHA: - fixed bugs of auto-configs of Spring Boot for MockMvc and WebTestClient
