--- neptune-spring-boot-starter-mockmvc/src/main/resources/META-INF/services/ru.tinkoff.qa.neptune.spring.boot.starter.mock.mvc.MockMvcProvider
@@ -1,1 +1,0 @@
-ru.tinkoff.qa.neptune.spring.boot.starter.mock.mvc.MockMvcProviderDefaultImpl
