--- neptune-spring-boot-starter-mockmvc/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=ru.tinkoff.qa.neptune.spring.boot.starter.mock.mvc.MockMvcWrappingConfiguration
