--- neptune-spring-boot-starter-webtest-client/README.md
@@ -1,27 +1,0 @@
-This module contains additional [auto-configurations](https://docs.spring.io/spring-boot/docs/2.1.13.RELEASE/reference/html/boot-features-developing-auto-configuration.html) 
-of  [Spring Boot](https://spring.io/guides/gs/spring-boot/) for the testing with [WebTestClient](./../spring.web.testclient/README.md).
-It is extension of [Neptune Spring Boot starter module](./../neptune-spring-boot-starter/README.md)
-
-
-## Maven
-
-```xml
-    <dependencies>
-        <dependency>
-            <groupId>ru.tinkoff.qa.neptune</groupId>
-            <artifactId>neptune-spring-boot-starter-webtest-client</artifactId>
-            <version>${LATEST_RELEASE_OR_BETA_VERSION}</version>
-            <scope>test</scope>
-        </dependency>
-    </dependencies>
-```
-
-## Gradle
-
-```groovy
-    dependencies {
-        testImplementation  group: 'ru.tinkoff.qa.neptune', name: 'neptune-spring-boot-starter-webtest-client', version: LATEST_RELEASE_OR_BETA_VERSION    
-    }
-```
-
-[API overview](https://tinkoffcreditsystems.github.io/neptune/neptune-spring-boot-starter-webtest-client/index.html)
