--- neptune-spring-boot-starter-webtest-client/src/main/resources/META-INF/services/ru.tinkoff.qa.neptune.spring.boot.starter.web.testclient.WebTestClientProvider
@@ -1,1 +1,0 @@
-ru.tinkoff.qa.neptune.spring.boot.starter.web.testclient.WebTestClientProviderDefaultImpl
