+++ neptune-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=ru.tinkoff.qa.neptune.spring.boot.starter.env.SpringEnvironmentWrappingConfiguration
 org.springframework.test.context.TestExecutionListener = ru.tinkoff.qa.neptune.spring.boot.starter.application.contexts.CurrentApplicationContextTestExecutionListener
